reportmem for the new stowedexception stuff
