+++ src/coreclr/src/debug/daccess/request.cpp
@@ -23,35 +23,6 @@
 #ifdef FEATURE_COMWRAPPERS
 #include <interoplibinterface.h>
 #include <interoplibabi.h>
-
-// Public contract for ExternalObjectContext, keep in sync with definition in
-// interoplibinterface.cpp
-struct ExternalObjectContextDACnterface
-{
-    PTR_VOID identity;
-    INT_PTR _padding1;
-    DWORD _padding2;
-    INT64 _padding3;
-};
-
-typedef DPTR(ExternalObjectContextDACnterface) PTR_ExternalObjectContext;
-
-// Public contract for ManagedObjectWrapper, keep in sync with definition in
-// comwrappers.hpp
-struct ManagedObjectWrapperDACInterface
-{
-    PTR_VOID managedObject;
-    INT32 _padding1;
-    INT32 _padding2;
-    INT_PTR _padding3;
-    INT_PTR _padding4;
-    INT_PTR _padding6;
-    LONGLONG _refCount;
-    INT32 _padding7;
-};
-
-typedef DPTR(ManagedObjectWrapperDACInterface) PTR_ManagedObjectWrapper;
-
 #endif // FEATURE_COMWRAPPERS
 
 #ifndef TARGET_UNIX
@@ -4104,22 +4075,23 @@ PTR_IUnknown ClrDataAccess::DACGetCOMIPFromCCW(PTR_ComCallWrapper pCCW, int vtab
 #endif
 
 #ifdef FEATURE_COMWRAPPERS
-BOOL ClrDataAccess::DACIsComWrappersCCW(CLRDATA_ADDRESS ccwPtr)
+BOOL ClrDataAccess::DACGetComWrappersCCWVTableQIAddress(CLRDATA_ADDRESS ccwPtr, TADDR *vTableAddress, TADDR *qiAddress)
 {
-    // Read CCWs QI address and compare it to the managed object wrapper's implementation.
+    _ASSERTE(vTableAddress != NULL && qiAddress != NULL);
+
     HRESULT hr = S_OK;
     ULONG32 bytesRead = 0;
     TADDR ccw = CLRDATA_ADDRESS_TO_TADDR(ccwPtr);
-    TADDR vTableAddress = NULL;
-    if (FAILED(m_pTarget->ReadVirtual(ccw, (PBYTE)&vTableAddress, sizeof(TADDR), &bytesRead))
+    *vTableAddress = NULL;
+    if (FAILED(m_pTarget->ReadVirtual(ccw, (PBYTE)vTableAddress, sizeof(TADDR), &bytesRead))
         || bytesRead != sizeof(TADDR)
         || vTableAddress == NULL)
     {
         return FALSE;
     }
 
-    TADDR qiAddress = NULL;
-    if (FAILED(m_pTarget->ReadVirtual(vTableAddress, (PBYTE)&qiAddress, sizeof(TADDR), &bytesRead))
+    *qiAddress = NULL;
+    if (FAILED(m_pTarget->ReadVirtual(*vTableAddress, (PBYTE)qiAddress, sizeof(TADDR), &bytesRead))
         || bytesRead != sizeof(TADDR)
         || qiAddress == NULL)
     {
@@ -4129,15 +4101,22 @@ BOOL ClrDataAccess::DACIsComWrappersCCW(CLRDATA_ADDRESS ccwPtr)
 
 #ifdef TARGET_ARM
     // clear the THUMB bit on qiAddress before comparing with known vtable entry
-    qiAddress &= ~THUMB_CODE;
+    *qiAddress &= ~THUMB_CODE;
 #endif
 
-    if (qiAddress != GetEEFuncEntryPoint(ManagedObjectWrapper_QueryInterface))
+    return TRUE;
+}
+
+BOOL ClrDataAccess::DACIsComWrappersCCW(CLRDATA_ADDRESS ccwPtr)
+{
+    TADDR vTableAddress = NULL;
+    TADDR qiAddress = NULL;
+    if (!DACGetComWrappersCCWVTableQIAddress(ccwPtr, &vTableAddress, &qiAddress))
     {
         return FALSE;
     }
 
-    return TRUE;
+    return qiAddress == GetEEFuncEntryPoint(ManagedObjectWrapper_QueryInterface);
 }
 
 TADDR ClrDataAccess::DACGetManagedObjectWrapperFromCCW(CLRDATA_ADDRESS ccwPtr)
@@ -4173,8 +4152,6 @@ HRESULT ClrDataAccess::DACTryGetComWrappersObjectFromCCW(CLRDATA_ADDRESS ccwPtr,
     }
 
     TADDR ccw = CLRDATA_ADDRESS_TO_TADDR(ccwPtr);
-    // Mask the "dispatch pointer" to get a double pointer to the ManagedObjectWrapper
-    TADDR managedObjectWrapperPtrPtr = ccw & InteropLib::ABI::DispatchThisPtrMask;
 
     // Return ManagedObjectWrapper as an OBJECTHANDLE. (The OBJECTHANDLE is guaranteed to live at offset 0).
     TADDR managedObjectWrapperPtr = DACGetManagedObjectWrapperFromCCW(ccwPtr);
